Add name to resolver
also restructure youtube resolver
