+++ internal/resolvers/seventv/emote_resolver.go
@@ -34,6 +34,10 @@ func (r *EmoteResolver) Run(ctx context.Context, url *url.URL, req *http.Request
 	return r.emoteCache.Get(ctx, emoteHash, req)
 }
 
+func (r *EmoteResolver) Name() string {
+	return "seventv:emote"
+}
+
 func NewEmoteResolver(ctx context.Context, cfg config.APIConfig) *EmoteResolver {
 	emoteLoader := &EmoteLoader{
 		baseURL: cfg.BaseURL,
