+++ pkg/resolver/resolver.go
@@ -9,4 +9,5 @@ import (
 type Resolver interface {
 	Check(ctx context.Context, url *url.URL) bool
 	Run(ctx context.Context, url *url.URL, r *http.Request) ([]byte, error)
+	Name() string
 }
